requirements.txt
anthropic~=0.62.0
python-dotenv~=1.1.1
pytest~=8.4.1